Refactoring: button folder
