+++ components/layout/ServiceLayout.tsx
@@ -2,7 +2,7 @@
 import { SWRConfig } from 'swr';
 import { ReactNode } from 'react';
 import Footer from '@/components/layout/Footer';
-import FloatButton from '@/components/common/FlotButton';
+import FloatButton from '@/components/common/button/FlotButton';
 import { ToastContainer } from 'react-toastify';
 import { PublicConfiguration } from 'swr/_internal';
 import Header from '@/components/layout/Header';
