+++ components/layout/sidebar/Sidebar.module.scss
@@ -35,7 +35,7 @@
     nav {
       display: flex;
       justify-content: space-between;
-      padding: calc(env(safe-area-inset-top)+ 0.5rem) 1rem 0 1rem;
+      padding: calc(env(safe-area-inset-top) + 0.5rem) 1rem 0 1rem;
       align-items: center;
 
       .close {
